add intersect point func
