+++ src/utils/intersects/intersect-point.js
@@ -1,0 +1,16 @@
+import { Polygon } from 'pixi.js';
+import { getViewport } from '../get';
+import { getPoints } from './get-points';
+
+export const intersectPoint = (obj, point) => {
+  const viewport = getViewport(obj);
+  if (!viewport) return false;
+
+  if ('containsPoint' in obj) {
+    return obj.containsPoint(point);
+  }
+
+  const points = getPoints(viewport, obj);
+  const polygon = new Polygon(points);
+  return polygon.contains(point.x, point.y);
+};
